+++ runtime/runtime.go
@@ -11,11 +11,11 @@ import (
 	"github.com/dotcloud/docker/runtime/execdriver/execdrivers"
 	"github.com/dotcloud/docker/runtime/execdriver/lxc"
 	"github.com/dotcloud/docker/graph"
-	"github.com/dotcloud/docker/graphdriver"
-	"github.com/dotcloud/docker/graphdriver/aufs"
-	_ "github.com/dotcloud/docker/graphdriver/btrfs"
-	_ "github.com/dotcloud/docker/graphdriver/devmapper"
-	_ "github.com/dotcloud/docker/graphdriver/vfs"
+	"github.com/dotcloud/docker/runtime/graphdriver"
+	"github.com/dotcloud/docker/runtime/graphdriver/aufs"
+	_ "github.com/dotcloud/docker/runtime/graphdriver/btrfs"
+	_ "github.com/dotcloud/docker/runtime/graphdriver/devmapper"
+	_ "github.com/dotcloud/docker/runtime/graphdriver/vfs"
 	"github.com/dotcloud/docker/image"
 	_ "github.com/dotcloud/docker/runtime/networkdriver/lxc"
 	"github.com/dotcloud/docker/runtime/networkdriver/portallocator"
